refactor(auth): name project seam files after their exports
